CODE/lineBot/StockAssit/requirements.txt
gunicorn
Flask
line-bot-sdk
pymongo
schedule
datetime
bs4
requests
clients
suds-py3
